app/requirements/requirements.txt
pillow==11